requirements.txt
fastapi
websockets==12.0
uvicorn==0.27.1
email-validator==2.1.0.post1
google-generativeai==0.4.1
python-dotenv==1.0.1
pandas-ta-classic
plotly
numpy
scikit-learn
tensorflow
transformers
torch
pyjwt
bcrypt
